-AviSynth now attempts to get handle from AviSynthpluginInit2 when loading plugins. -internal.h can now be left out of plugin development.

+++ avisynth.h
@@ -20,6 +20,9 @@
 #ifndef __AVISYNTH_H__
 #define __AVISYNTH_H__
 
+#include <crtdbg.h>
+#include <objbase.h>
+
 
 enum { AVISYNTH_INTERFACE_VERSION = 2 };
 
@@ -117,7 +120,7 @@ struct VideoInfo {
   int RowSize() const { return BytesFromPixels(width); }  // Also only returns first plane on planar images
   int BMPSize() const { if (IsPlanar()) {int p = height * ((RowSize()+3) & ~3); p+=p>>1; return p;  } return height * ((RowSize()+3) & ~3); }
   __int64 AudioSamplesFromFrames(__int64 frames) const { return (__int64(frames) * audio_samples_per_second * fps_denominator / fps_numerator); }
-  int FramesFromAudioSamples(__int64 samples) const { return (__int64(samples) * fps_numerator / fps_denominator / audio_samples_per_second); }
+  int FramesFromAudioSamples(__int64 samples) const { return (int)(samples * (__int64)fps_numerator / (__int64)fps_denominator / (__int64)audio_samples_per_second); }
   __int64 AudioSamplesFromBytes(__int64 bytes) const { return bytes / BytesPerAudioSample(); }
   __int64 BytesFromAudioSamples(__int64 samples) const { return samples * BytesPerAudioSample(); }
   int AudioChannels() const { return nchannels; }
